restructure project packages
ristrutturazione della gerarchia delle cartelle che compongono il progetti
