+++ src/main/java/com/example/capstone/arkadia/libris/dto/request/UpdateUserDto.java
@@ -1,4 +1,4 @@
-package com.example.capstone.arkadia.libris.dto;
+package com.example.capstone.arkadia.libris.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
