+++ src/main/java/com/example/capstone/arkadia/libris/repository/AddressRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Addressrepository extends JpaRepository<Address, Long> {}
+public interface AddressRepository extends JpaRepository<Address, Long> {}
