+++ src/main/java/com/example/capstone/arkadia/libris/security/JwtTool.java
@@ -2,7 +2,7 @@ package com.example.capstone.arkadia.libris.security;
 
 import com.example.capstone.arkadia.libris.exception.NotFoundException;
 import com.example.capstone.arkadia.libris.model.User;
-import com.example.capstone.arkadia.libris.service.UserService;
+import com.example.capstone.arkadia.libris.service.user.UserService;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Autowired;
